# fs: allocate backing storage once in Fs::load